backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
pydantic
PyJWT
passlib[bcrypt]
pillow
pytesseract
requests